// models/sales_paper_bag.model.ts
export type SalesPaperBagState = {
  data: {
    sals_papb_id: string;
    sals_papb_date: string;
    sals_papb_fullname: string;
    sals_papb_company_name: string;
    sals_papb_tel: string;
    sals_papb_email: string;
    sals_papb_doc_type: string;
    sals_papb_printing_type: string;
    sals_papb_amount: string;
    sals_papb_quotation_request: string;
    sals_papb_finished_size: string;
    sals_papb_page: string;
    sals_papb_paper: string;
    sals_papb_printing: string;
    sals_papb_coating: string;
    sals_papb_binding: string;
    sals_papb_printing_volume: string;
  };
};